BLEN-257: Make GL delegate to be disabled in USDHydra render settings. Preparations, refactoring
