+++ extern/usdhydra/addon/__init__.py
@@ -19,29 +19,22 @@ bl_info = {
     "category": "Render"
 }
 
-# Support 'reload' case.
-if "bpy" in locals():
-    import importlib
-    importlib.reload(properties)
-    importlib.reload(engine)
-    importlib.reload(ui)
-    # importlib.reload(operators)
-
-else:
-    from . import (
-        properties,
-        engine,
-        usd_nodes,
-        ui,
-        handlers,
-    )
-
 import atexit
+
 import bpy
 from bpy.utils import register_class, unregister_class
-from .utils import logging
 
+from .utils import logging
 log = logging.Log('init')
+
+from . import (
+    properties,
+    engine,
+    usd_nodes,
+    ui,
+    handlers,
+    preferences,
+)
 
 
 def exit():
@@ -59,7 +52,7 @@ def register():
     usd_nodes.register()
     handlers.register()
 
-    register_class(engine.USDHydraHdStormEngine)
+    register_class(preferences.AddonPreferences)
 
     from .utils import enable_delegates
     enable_delegates()
@@ -73,4 +66,4 @@ def unregister():
     usd_nodes.unregister()
     handlers.unregister()
 
-    unregister_class(engine.USDHydraHdStormEngine)
+    unregister_class(preferences.AddonPreferences)
